+++ CHANGELOG.md
@@ -54,6 +54,10 @@ and this project adheres to
 - A Unicode character used as `CRLF-CHARACTER` no longer gets added to the
   Unicode translation table.
 
+- The abbreviation finder now correctly counts substrings that overlap with
+  themselves (e.g. repeated characters like "------"), resulting in better
+  abbreviation choices in some cases.
+
 ### Removed
 
 - Removed `dragon` sample. Those wishing to fight dragons are encouraged to try